fix(nocodb): fail closed on malformed filterArrJson across all count paths
