+++ packages/nocodb/src/dbQueryClient/cross-db-utils/bulk-aggregate.ts
@@ -6,43 +6,9 @@ import type { BulkAggregateCtx, DBQueryClient } from '~/dbQueryClient/types';
 import { applyAggregation } from '~/dbQueryClient/cross-db-utils/applyAggregation';
 import conditionV2 from '~/db/conditionV2';
 import { Filter, Model } from '~/models';
-import { NcError } from '~/helpers/ncError';
+import { parseFilterArrJson } from '~/helpers/filterArrJsonHelper';
 import NcConnectionMgrv2 from '~/utils/common/NcConnectionMgrv2';
 import { resolveAggregateColumns } from '~/dbQueryClient/cross-db-utils/aggregate';
-
-// filterArrJson arrives either as a JSON string (API callers) or an
-// already-parsed Filter[] (internal callers pass it pre-parsed). A malformed
-// JSON string must fail closed with a 400 — silently dropping it would run the
-// aggregation unfiltered (fail-open data exposure). Empty/absent => no extra
-// filter; a value that parses to a non-array is rejected too.
-function parseBulkFilterArrJson(
-  context: NcContext,
-  raw: string | Filter[] | undefined,
-  alias: string,
-): Filter[] | undefined {
-  if (raw == null) return undefined;
-  if (Array.isArray(raw)) return raw;
-
-  const trimmed = raw.trim();
-  if (!trimmed) return undefined;
-
-  let parsed: unknown;
-  try {
-    parsed = JSON.parse(trimmed);
-  } catch {
-    NcError.get(context).badRequest(
-      `Invalid filterArrJson for bulk-aggregate bucket "${alias}"`,
-    );
-  }
-
-  if (!Array.isArray(parsed)) {
-    NcError.get(context).badRequest(
-      `Invalid filterArrJson for bulk-aggregate bucket "${alias}"`,
-    );
-  }
-
-  return parsed as Filter[];
-}
 
 /**
  * Shared, dialect-agnostic bulk aggregation orchestration.
@@ -71,7 +37,11 @@ export const bulkAggregate =
     const parsedFilterArrJsonByAlias = new Map(
       (bulkFilterList ?? []).map((f): [string, Filter[] | undefined] => [
         f.alias,
-        parseBulkFilterArrJson(context, f.filterArrJson, f.alias),
+        parseFilterArrJson(
+          context,
+          f.filterArrJson,
+          `bulk-aggregate bucket "${f.alias}"`,
+        ),
       ]),
     );
